homepage and add function
