+++ nav.php
@@ -1,7 +1,7 @@
 <ul class="navbar-nav bg-gradient-primary sidebar sidebar-dark accordion" id="accordionSidebar">
 
     <!-- Sidebar - Brand -->
-    <a class="sidebar-brand d-flex align-items-center justify-content-center" href="index.php">
+    <a class="sidebar-brand d-flex align-items-center justify-content-center" href="homepage.php">
         <div class="sidebar-brand-icon rotate-n-15">
             <i class="fas fa-location-arrow"></i>
         </div>
@@ -13,7 +13,7 @@
 
     <!-- Nav Item - Dashboard -->
     <li id="dashboard_ad" class="nav-item active">
-        <a class="nav-link" href="index.php">
+        <a class="nav-link" href="homepage.php">
             <i class="fas fa-fw fa-tachometer-alt"></i>
             <span>Dashboard</span></a>
     </li>
